build(content-grid): fix content and new test build
