+++ @uportal/content-grid/package.json
@@ -1,34 +1,45 @@
 {
   "name": "content-grid",
   "version": "0.1.0",
-  "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build --name content-grid --target wc --mode production",
+    "buildold": "vue-cli-service build --name content-grid --target wc 'src/components/ContentGrid.vue'",
+    "build": "vue-cli-service build --name content-grid --target wc 'src/mainwc.js'",
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@uportal/open-id-connect": "^1.1.0",
+    "@uportal/open-id-connect": "^1.2.0",
+    "portlet-card": "^0.1.7",
     "vue": "^2.5.16",
-    "vue-i18n": "^7.8.1",
-    "vue-line-clamp": "^1.2.3"
+    "vue-i18n": "^7.8.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^3.0.0-rc.3",
     "@vue/cli-plugin-eslint": "^3.0.0-rc.3",
     "@vue/cli-service": "^3.0.0-rc.3",
     "@vue/eslint-config-prettier": "^3.0.0-rc.3",
-    "css-loader": "^0.28.11",
     "node-sass": "^4.9.0",
     "sass-loader": "^7.0.3",
     "vue-template-compiler": "^2.5.16"
   },
-  "entry": [
-    "babel-polyfill",
-    "./src/main.js"
-  ],
-  "vue": {
-    "lintOnSave": true
+  "babel": {
+    "presets": [
+      "@vue/app"
+    ]
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/essential",
+      "eslint:recommended"
+    ],
+    "rules": {},
+    "parserOptions": {
+      "parser": "babel-eslint"
+    }
   },
   "postcss": {
     "plugins": {
@@ -37,9 +48,28 @@
   },
   "browserslist": [
     "> 1%",
-    "not ie < 11"
+    "last 2 versions",
+    "not ie <= 8"
   ],
   "publishConfig": {
     "access": "public"
-  }
+  },
+  "main": "dist/content-grid.js",
+  "description": "Display all browsable uPortal portlet as cards",
+  "license": "Apache-2.0",
+  "author": "Julien Gribonvald <julien.gribonvald@gmail.com>",
+  "keywords": [
+    "apereo",
+    "uportal",
+    "portlet",
+    "grid"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/jgribonvald/uPortal-web-components.git"
+  },
+  "bugs": {
+    "url": "https://github.com/jgribonvald/uPortal-web-components/issues"
+  },
+  "homepage": "https://github.com/jgribonvald/uPortal-web-components#readme"
 }
