+++ @uportal/content-grid/src/mainwc.js
@@ -1,0 +1,7 @@
+import Vue from "vue";
+import wrap from "@vue/web-component-wrapper";
+import contentGrid from "./components/ContentGrid.vue";
+
+const CustomElement = wrap(Vue, contentGrid);
+
+window.customElements.define("content-grid", CustomElement);
